requirements.txt
flask==3.0.0
flask-cors==4.0.0
langchain==0.1.10
langchain-community==0.0.29
langchain-google-genai==1.0.1
google-generativeai==0.4.1
faiss-cpu==1.8.0
python-dotenv==1.0.1
gunicorn==21.2.0
sentence-transformers==2.2.2